Banned users, Private audio

+++ src/rest/user/Admin.java
@@ -3,10 +3,12 @@ package rest.user;
 import BusinessLogic.UserLogic;
 import DAO.UserDAO;
 import DAO.util.Factory;
+import Entity.BannedEntity;
 import Entity.UserEntity;
 import org.apache.log4j.Logger;
 import util.EmailSender;
 
+import javax.mail.MessagingException;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -66,20 +68,27 @@ public class Admin {
                     log.info("ADMIN: " + admin.getLogin() + " user not found... \'in ban\'");
                     return Response.ok().status(400).build();
                 }
-                user.setPassword(UserLogic.uid(40));
+                if (user.getId() == admin.getId()) {
+                    log.info("ADMIN: " + admin.getLogin() + " you cant banned yourself");
+                    return Response.ok().status(403).build();
+                }
+                if (user.getAccess() > 0) {
+                    log.info("ADMIN: " + admin.getLogin() + " you cant banned moderator or admin");
+                    return Response.ok().status(404).build();
+                }
+                Factory.getInstance().getBannedDAO().add(new BannedEntity(user.getId()));
                 String email = user.getEmail();
-                userDAO.change(user);
                 EmailSender.generateAndSendEmail(email, "You are banned, motherfucker!");
                 Factory.getInstance().getSessionDAO().delete(userID);
                 log.info("ADMIN: " + admin.getLogin() + " success banned user: " + user.getLogin());
                 return Response.ok().status(200).build();
             } catch (Exception e) {
                 log.info("ADMIN: " + admin.getLogin() + " have some trouble with banned user");
-                return Response.ok().status(203).build();
+                return Response.ok().status(401).build();
             }
         } else {
             log.info("ADMIN: was not admin... \'in ban\'");
-            return Response.ok().status(204).build();
+            return Response.ok().status(402).build();
         }
     }
 
@@ -91,26 +100,36 @@ public class Admin {
     ) {
         UserDAO userDAO = Factory.getInstance().getUserDAO();
         UserEntity admin = userDAO.getById(Factory.getInstance().getSessionDAO().haveKey(uid));
+        if(admin == null)
+            return Response.status(400).entity("logged in pls").build();
         if (admin.getAccess() == 2) {
             try {
                 UserEntity user = userDAO.getById(userID);
                 if (user == null) {
                     log.info("ADMIN: " + admin.getLogin() + " user not found... \'in unban\'");
-                    return Response.ok().status(400).build();
+                    return Response.ok().status(400).entity("user not found").build();
                 }
-                String email = user.getEmail();
-                String pass = user.getPassword();
-                String message = "Congratz, you are unbanned ! Here is your new password: " + pass;
-                EmailSender.generateAndSendEmail(email, message);
-                log.info("ADMIN: " + admin.getLogin() + " success unbanned user: " + user.getLogin());
-                return Response.ok().status(200).build();
+                if(Factory.getInstance().getBannedDAO().isUserId(userID)) {
+                    Factory.getInstance().getBannedDAO().delete(user);
+                    String email = user.getEmail();
+                    String message = "Congratz, you are unbanne!";
+                    try {
+                        EmailSender.generateAndSendEmail(email, message);
+                    } catch (MessagingException exception){}
+
+                    log.info("ADMIN: " + admin.getLogin() + " success unbanned user: " + user.getLogin());
+                    return Response.ok().status(200).build();
+                } else {
+                    log.info("ADMIN: " + admin.getLogin() + " user: " + user.getLogin()+" not in banlist");
+                    return Response.ok().status(400).entity("user not banned").build();
+                }
             } catch (Exception e) {
                 log.info("ADMIN: " + admin.getLogin() + " have some trouble with banned user");
-                return Response.ok().status(400).build();
+                return Response.ok().status(400).entity("some problem").build();
             }
         } else {
             log.info("ADMIN: was not admin... \'in unban\'");
-            return Response.ok().status(204).build();
+            return Response.ok().status(404).entity("was not admin").build();
         }
     }
 }
